More organized code for js page
