+++ client/src/pages/Notes/JavaScriptFundamentals/JsPageTitleManager.jsx
